combined distribution chart component
